requirements.txt
Flask==2.3.2
requests==2.32.3
pandas==2.2.3
sendgrid==6.11.0
python-dotenv==1.0.0
gunicorn==21.2.0